train/requirements.txt
absl-py==2.1.0
accelerate==0.29.2
addict==2.4.0
adjustText==1.1.1
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpathlib==0.16.0
cmake==3.29.0.1
-e git+https://github.com/facebookresearch/co3d.git@ff54df4dfcf5ac3d653d653b5a32b8e5653c70a9#egg=co3d
colorama==0.4.6
comm==0.2.2
confection==0.1.4
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.24.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.4
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
ftfy==6.1.3
fvcore==0.1.5.post20221221
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gradio==4.24.0
gradio_client==0.14.0
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.0.0
install==1.3.5
iopath==0.1.10
ipyfilechooser==0.6.0
ipykernel==6.29.4
ipympl==0.9.3
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.2
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
kmeans-pytorch @ git+https://github.com/subhadarship/kmeans_pytorch.git@f7f36bd1cb4e3a761d73d584866d0a9c6b4d2805
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
lit==18.1.2
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.10.2
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.59.0
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.0
overrides==7.4.0
packaging==24.0
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.0
plotly==5.20.0
plyfile==1.0.3
portalocker==2.8.2
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pyquaternion==0.9.9
pyre-extensions==0.0.29
pySankeyBeta==1.4.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
ruff==0.3.4
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.18.0
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
timm @ git+https://github.com/huggingface/pytorch-image-models.git@3df4ffe914f7969344fc728c0b8a9b8289b3baa9
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torch-summary==1.4.5
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.36.2
triton==2.0.0
typer==0.9.4
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
visdom==0.2.4
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xformers==0.0.20
yacs==0.1.8
zipp==3.18.1
